requirements.txt
feedparser==6.0.10
requests==2.31.0
beautifulsoup4==4.12.2
python-dateutil==2.8.2
urllib3==2.1.0

# Testing
pytest==7.4.4
pytest-cov==4.1.0
